chg: multiple changes to the qsim simulator.
