+++ setup.cfg
@@ -35,6 +35,8 @@ install_requires =
 where = src
 
 [options.extras_require]
+qsim =
+    qsimcirq>=0.12.1
 dev =
     orquestra-python-dev
