+++ src/orquestra/integrations/cirq/simulator/__init__.py
@@ -1,4 +1,9 @@
 ################################################################################
 # © Copyright 2021-2022 Zapata Computing Inc.
 ################################################################################
-from .simulator import CirqSimulator, QsimSimulator
+from .simulator import CirqSimulator
+
+try:
+    from .qsim_simulator import QsimSimulator
+except:
+    pass
